requirements.txt
sqlalchemy==1.4.32
fastapi==0.75.0
pydantic==1.9.0
jinja2==3.0.3
asyncpg==0.25.0
uvicorn==0.17.6
pyjwt==2.3.0